property_rag_clean/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
streamlit==1.31.0
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3